discord-bot/requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
pytz>=2024.1
